requirements.txt
qdrant-client
pandas
numpy
streamlit
scikit-learn
yellowbrick
matplotlib
faiss-cpu
sentence-transformers
rank-bm25
langdetect
deep-translator
seaborn
